pyproject.toml
[tool.poetry]
name = "calico"
version = "0.1.0"
description = ""
authors = ["Jacob Monck <jacobamonck@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.3.0"
python-dotenv = "^0.19.2"
loguru = "^0.5.3"
tortoise-orm = "^0.18.0"
asyncpg = "^0.25.0"
PyYAML = "^6.0"
taskipy = "^1.9.0"
more-itertools = "^8.12.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"

[tool.taskipy.tasks]
format = "black . && isort --profile black ."
bot = "python3 -m src"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
